requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
openai>=1.0.0
python-dotenv>=1.0.0